+++ CHANGES.md
@@ -43,6 +43,7 @@
 - Fix decimal numbers in export viewbox [Taiga #2290](https://tree.taiga.io/project/penpot/issue/2290)
 - Right click over artboard name to open its menu [Taiga #1679](https://tree.taiga.io/project/penpot/issue/1679)
 - Make the default session cookue use SameSite=Lax instead of Strict (causes some issues in latest versions of Chrome).
+- Fix "open in new tab" on dashboard [Taiga #2235](https://tree.taiga.io/project/penpot/issue/2355)
 
 ### :arrow_up: Deps updates